+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- Fixed a bug where `\craft\filters\Headers` and `\craft\filters\Cors` were applied to CP requests instead of site requests.
+- Fixed a bug where `\craft\filters\Headers` and `\craft\filters\Cors` were applied to control panel requests instead of site requests.
 
 ## 4.11.0.2 - 2024-08-06